+++ src/app/features/credential-management/credential-management.component.ts
@@ -109,8 +109,6 @@ export class CredentialManagementComponent implements OnInit, AfterViewInit {
   public ngOnInit() {
     this.loadCredentialData();
     this.isAdminOrganizationIdentifier = this.authService.hasAdminOrganizationIdentifier();
-    console.log("Management: isAdminOrganizationIdentifier");
-    console.log(this.isAdminOrganizationIdentifier);
   }
 
   public ngAfterViewInit(): void {